Moved Tempate Object form to Obsolete. Moved compoents to Schema Document form. Removed any remaining refrences.
