+++ DataDictionary.Main/Forms/Obsolete/TemplateObject.Binding.cs
@@ -3,7 +3,7 @@ using DataDictionary.BusinessLayer.NamedScope;
 using DataDictionary.BusinessLayer.ToolSet;
 using System.Data;
 
-namespace DataDictionary.Main.Forms.Scripting
+namespace DataDictionary.Main.Forms.Obsolete
 {
     partial class TemplateObject
     {
