+++ DataDictionary.Main/Forms/Obsolete/TemplateObject.resx
