+++ DataDictionary.Main/Forms/Scripting/SchemaDocument.resx
@@ -123,7 +123,27 @@
   <metadata name="schemaDocumentLayout.GenerateMember" type="System.Boolean, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089">
     <value>False</value>
   </metadata>
+  <metadata name="objectGroupBox.GenerateMember" type="System.Boolean, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089">
+    <value>False</value>
+  </metadata>
+  <metadata name="objectLayout.GenerateMember" type="System.Boolean, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089">
+    <value>False</value>
+  </metadata>
+  <metadata name="objectBehaviorLayout.GenerateMember" type="System.Boolean, mscorlib, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089">
+    <value>False</value>
+  </metadata>
   <assembly alias="System.Drawing" name="System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a" />
+  <data name="objectNameData.SelectIcon" type="System.Drawing.Bitmap, System.Drawing" mimetype="application/x-microsoft.net.object.bytearray.base64">
+    <value>
+        iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAABGdBTUEAALGPC/xhBQAAACBjSFJNAAB6
+        JgAAgIQAAPoAAACA6AAAdTAAAOpgAAA6mAAAF3CculE8AAAACXBIWXMAAA7DAAAOwwHHb6hkAAABDUlE
+        QVQ4T8WQsUoEMRRFTyX4CaJ/YOMHWIh+gljtRv0BX0bZzn7Rwm7ZTaaz1FqEZG0FKztbFUVhrbSwXplZ
+        Rs3TgILghUNCJveQefDbtGOH1nBJH/88ZriIiSPaYVl/ymfteAoT9zHxERPHE8KVvpaPCXuYcML62Wz9
+        /HZ8Sn7DCbfeMv7Ey8DyccGEB1qnc/X+uxnUpYKZhr6wOhBeq/OJINarF26SYhMtyNIIdf5EUHZY8JZn
+        NYs8QkgETuh6y1E1vBx9S1GtvuDQC71E4IW70rKpS1rQ22HFC6OyYD4VWC51QVO/YIvdUrh4LzcCV9DV
+        BU0lcMK5FzYSgbNcfxlSBme5P9hmOhH8a94ANSDaAJQg0nQAAAAASUVORK5CYII=
+</value>
+  </data>
   <data name="documentFileData.SelectIcon" type="System.Drawing.Bitmap, System.Drawing" mimetype="application/x-microsoft.net.object.bytearray.base64">
     <value>
         iVBORw0KGgoAAAANSUhEUgAAABAAAAAQCAYAAAAf8/9hAAAABGdBTUEAALGPC/xhBQAAACBjSFJNAAB6
